Adding new logo (#526)
* Adding new logo

I reverted to the previous way of displaying the logo, as a background
SVG image. Adding a new `<svg>` inline element to the markup could
break people implementation relying on our markup, so better play safe
and keep the markup the same.

I used the SVG available on our
https://www.algolia.com/press#resources page, but trimmed it through
SVGO (50% compression) before adding it to the background.

* Updating logo to version validated by design team

Original dimensions are 168x24, but this is substancially bigger than what we
used to have (130x19px). InstantSearch used a 134x20px ratio, and
I re-used the same. Slightly bigger than what we had before, but
keeping the correct ratio.