Allow release from `yarn run`
Running `yarn run release` causes issues as it can't communicate with
the proper npm registry. This happens because yarn overrides the
`npm_config_registry` variable, on which the publishing depends.

In this commit I revert the value to its default, allowing the release
script to be run either with `npm run release` or `yarn run release`.