docs(deploy-preview): Re-enable deploy previews for the doc
This re-enable deploy-previews through Netlify. It will check if
changes where made to the ./docs folder and build + deploy the preview
if so. Otherwise, it will deploy a dummy preview.

This also removes the netlify script to deploy in production whenever
we push to master, as this is now handled by Travis.