fix: updating the templates page (#2273)
* fix: updating the templates page

The templates page needs some work. The first thing I saw is that it alluded to being able to create a crawler. As a docsearch user, this is confusing. We automatically create a crawler for the user, additional crawlers get confusing for the user.

Provide additional information about updating templates, as we're talking about templates here.

Verify the current templates, I see the default one isn't quite right.

Given that we use the same template for Docusaurus v2 and v3 combine those.

Add the Starlight template.

* Combined Docusaurus v2 and v3 templates

Its confusing to have two separate sections that are exactly the same

* added astro starlight template

* remove pkgdown and add starlight

pkgdown no longer references Algolia
added starlight as an integration partner