fix: size and background scroll fixes (#2603)
* fix(docsearch-css): use correct size and position

`position: absolute` makes the element absolute relative to its parent
If the parent is not sized/positioned as desired, then the children
suffer. `position: fixed` makes this element no longer depend on parent
positioning.

The height units are changed to more reliable/modern units to get the
desired behavior.

* fix(docksearch-css): prevent mobile body scrolling

On initial modal open, the body is scrollable. Prevent this as suggested
on SO

---------

Co-authored-by: Dylan Tientcheu <dylantientcheu@gmail.com>